products display and categories display improvement

+++ admin-app/src/actions/authAction.js
@@ -67,9 +67,22 @@ export const isUserLoggedIn=()=>{
 
 export const LogOut=()=>{
     return async dispatch =>{
-        localStorage.clear()
-        dispatch({
-            type:authConstant.LOGOUT_REQUEST
-        })
+
+        dispatch({type:authConstant.LOGOUT_REQUEST})
+        const res = await axiosInstance.post('./admin/signout')
+
+        if(res.status===200){
+            localStorage.clear()
+            dispatch({
+                type:authConstant.LOGOUT_SUCCESS
+            })
+        }
+        else{
+            dispatch({
+                type:authConstant.LOGOUT_FAILURE,
+                payload:{error:res.data.error}
+            })
+        }
+        
     }
 }
